#2162 Avoid double input in data point properties edition: - Corrected handler exception; - Moved EngineeringUnitJson to org.scada_lts.web.mvc.api package;
